requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.28.0
datasets==2.15.0
pandas==2.1.0
numpy==1.26.0
scikit-learn==1.3.2
accelerate==0.25.0
peft==0.6.2